requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
plotly==5.22.0
kaleido==0.2.1
python-dotenv==1.0.1
